Tests/Tokenizer: move tests
Move the tests directly related to the `Tokenizer\PHP` class into a `PHP` subdirectory.
Move the tests directly related to the `Tokenizer\Tokenizer` class into a `Tokenizer` subdirectory.

As there are five (soon three) Tokenizer classes, having the tests for each class in their own subdirectory makes it more straight forward to see what code is covered by tests.
