+++ tests/Core/Tokenizer/PHP/AnonClassParenthesisOwnerTest.inc
