+++ tests/Core/Tokenizer/PHP/AttributesTest.inc
