+++ tests/Core/Tokenizer/PHP/BackfillExplicitOctalNotationTest.php
@@ -7,7 +7,9 @@
  * @license   https://github.com/PHPCSStandards/PHP_CodeSniffer/blob/master/licence.txt BSD Licence
  */
 
-namespace PHP_CodeSniffer\Tests\Core\Tokenizer;
+namespace PHP_CodeSniffer\Tests\Core\Tokenizer\PHP;
+
+use PHP_CodeSniffer\Tests\Core\Tokenizer\AbstractTokenizerTestCase;
 
 final class BackfillExplicitOctalNotationTest extends AbstractTokenizerTestCase
 {
