+++ tests/Core/Tokenizer/PHP/BitwiseOrTest.inc
