+++ tests/Core/Tokenizer/PHP/ContextSensitiveKeywordsTest.php
@@ -7,8 +7,9 @@
  * @license   https://github.com/PHPCSStandards/PHP_CodeSniffer/blob/master/licence.txt BSD Licence
  */
 
-namespace PHP_CodeSniffer\Tests\Core\Tokenizer;
+namespace PHP_CodeSniffer\Tests\Core\Tokenizer\PHP;
 
+use PHP_CodeSniffer\Tests\Core\Tokenizer\AbstractTokenizerTestCase;
 use PHP_CodeSniffer\Util\Tokens;
 
 final class ContextSensitiveKeywordsTest extends AbstractTokenizerTestCase
