+++ tests/Core/Tokenizer/PHP/DoubleArrowTest.inc
