+++ tests/Core/Tokenizer/PHP/DoubleQuotedStringTest.inc
