+++ tests/Core/Tokenizer/PHP/FinallyTest.inc
