+++ tests/Core/Tokenizer/PHP/GotoLabelTest.inc
