+++ tests/Core/Tokenizer/PHP/HeredocStringTest.inc
