+++ tests/Core/Tokenizer/PHP/NullsafeObjectOperatorTest.inc
