+++ tests/Core/Tokenizer/PHP/ResolveSimpleTokenTest.inc
