+++ tests/Core/Tokenizer/PHP/ShortArrayTest.inc
