+++ tests/Core/Tokenizer/PHP/TypeIntersectionTest.inc
