+++ tests/Core/Tokenizer/PHP/TypedConstantsTest.inc
