+++ tests/Core/Tokenizer/PHP/UndoNamespacedNameSingleTokenTest.inc
