+++ tests/Core/Tokenizer/Tokenizer/HeredocNowdocCloserTest.inc
