+++ tests/Core/Tokenizer/Tokenizer/ScopeSettingWithNamespaceOperatorTest.inc
